+++ e2e/cypress/dataHooks.js
@@ -3,6 +3,7 @@ export const INLINE_TOOLBAR_BUTTONS = {
   ITALIC: 'textInlineStyleButton_Italic',
   UNDERLINE: 'textInlineStyleButton_Underline',
   COLOR: 'TextColorButton',
+  HIGHTLIGHT: 'TextHighlightButton',
   TITLE: 'textBlockStyleButton_Title',
   QUOTE: 'textBlockStyleButton_Quote',
   DECREASE_INDENT: 'decreaseIndentButton',
@@ -32,6 +33,7 @@ export const STATIC_TOOLBAR_BUTTONS = {
   BOOKING: 'Bookings_InsertButton',
   GIPHY: 'GIFPlugin_InsertButton',
   ADSENSE: 'AdSensePlugin_InsertButton',
+  EMOJI: 'EmojiPlugin_InsertButton',
 };
 
 export const SETTINGS_PANEL = {