+++ e2e/cypress/support/commands.js
@@ -1,7 +1,5 @@
 /*global Cypress, cy*/
 require('cypress-plugin-snapshots/commands');
-import { addMatchImageSnapshotCommand } from 'cypress-image-snapshot/command';
-addMatchImageSnapshotCommand();
 import {
   INLINE_TOOLBAR_BUTTONS,
   PLUGIN_TOOLBAR_BUTTONS,
@@ -118,11 +116,37 @@ Cypress.Commands.add('matchContentSnapshot', () => {
       .toMatchSnapshot();
 });
 
-Cypress.Commands.add('matchSnapshots', options => {
-  cy.matchImageSnapshot(options).matchContentSnapshot();
+Cypress.Commands.add('getViewer', () => {
+  cy.get('[data-hook="ricos-viewer"]');
+});
+
+Cypress.Commands.add('getTwitterButton', () => {
+  cy.get('[data-hook="twitter-button"]');
+});
+
+function setSelection(start, offset, container) {
+  container.then(args => {
+    const getTextElmentAndLocalOffset = getTextElments(args[0]);
+    const document = args[0].ownerDocument;
+    const range = document.createRange();
+    const startObj = getTextElmentAndLocalOffset(start);
+    range.setStart(startObj.element, startObj.offset);
+    const endObj = getTextElmentAndLocalOffset(start + offset);
+    range.setEnd(endObj.element, endObj.offset);
+    document.getSelection().removeAllRanges(range);
+    document.getSelection().addRange(range);
+  });
+}
+
+Cypress.Commands.add('setViewerSelection', (start, offset) => {
+  setSelection(start, offset, cy.getViewer());
 });
 
 // Editor commands
+
+Cypress.Commands.add('setEditorSelection', (start, offset) => {
+  setSelection(start, offset, cy.focusEditor());
+});
 
 Cypress.Commands.add('enterText', text => {
   cy.getEditor().type(text);
@@ -214,20 +238,6 @@ function getTextElments(rootElement) {
   };
 }
 
-Cypress.Commands.add('setSelection', (start, offset) => {
-  cy.focusEditor().then(args => {
-    const getTextElmentAndLocalOffset = getTextElments(args[0]);
-    const document = args[0].ownerDocument;
-    const range = document.createRange();
-    const startObj = getTextElmentAndLocalOffset(start);
-    range.setStart(startObj.element, startObj.offset);
-    const endObj = getTextElmentAndLocalOffset(start + offset);
-    range.setEnd(endObj.element, endObj.offset);
-    document.getSelection().removeAllRanges(range);
-    document.getSelection().addRange(range);
-  });
-});
-
 Cypress.Commands.add('moveCursorToStart', () => {
   cy.focusEditor().type('{selectall}{uparrow}');
 });
@@ -238,9 +248,21 @@ Cypress.Commands.add('moveCursorToEnd', () => {
 
 Cypress.Commands.add('setTextStyle', (buttonSelector, selection) => {
   if (selection) {
-    cy.setSelection(selection[0], selection[1]);
+    cy.setEditorSelection(selection[0], selection[1]);
   }
-  cy.get(`[data-hook=inlineToolbar] [data-hook=${buttonSelector}]`).click();
+  cy.get(
+    `[data-hook=${isMobile ? 'mobileToolbar' : 'inlineToolbar'}] [data-hook=${buttonSelector}]`
+  ).click();
+});
+
+Cypress.Commands.add('setTextColor', (selection, color) => {
+  cy.setTextStyle(INLINE_TOOLBAR_BUTTONS.COLOR, selection);
+  cy.get(`[data-scheme-color="${color}"]`).click();
+});
+
+Cypress.Commands.add('setHighlightColor', (selection, color) => {
+  cy.setTextStyle(INLINE_TOOLBAR_BUTTONS.HIGHTLIGHT, selection);
+  cy.get(`[data-scheme-color="${color}"]`).click();
 });
 
 Cypress.Commands.add('increaseIndent', selection => {
@@ -295,6 +317,11 @@ Cypress.Commands.add('openSideToolbar', () => {
 
 Cypress.Commands.add('openAddPluginModal', () => {
   cy.get('[data-hook="addPluginButton"]').click();
+  cy.get('[data-hook="addPluginMenu"]');
+});
+
+Cypress.Commands.add('openFooterPluginMenu', () => {
+  cy.get('[data-hook="moreButton"]').click();
   cy.get('[data-hook="addPluginMenu"]');
 });
 
@@ -405,6 +432,13 @@ Cypress.Commands.add('addImageLink', () => {
   // .get('href=www.wix.com');
 });
 
+Cypress.Commands.add('getImageLink', () => {
+  cy.openPluginToolbar(PLUGIN_COMPONENT.IMAGE)
+    .clickToolbarButton(PLUGIN_TOOLBAR_BUTTONS.LINK)
+    .get(`[data-hook=linkPanelContainer] [data-hook=linkPanelInput]`)
+    .should('have.value', 'www.wix.com');
+});
+
 Cypress.Commands.add('alignImage', alignment => {
   let button;
   switch (alignment) {
@@ -480,6 +514,10 @@ Cypress.Commands.add('addVideoFromURL', () => {
 
 Cypress.Commands.add('clickOnStaticButton', dataHook =>
   cy.get(`[data-hook*=footerToolbar] [data-hook*=${dataHook}]`).click()
+);
+
+Cypress.Commands.add('clickOnPluginMenuButton', dataHook =>
+  cy.get(`[data-hook*=addPluginMenu] [data-hook*=${dataHook}]`).click({force: true})
 );
 
 Cypress.Commands.add('addHtml', () => {
@@ -604,9 +642,3 @@ Cypress.Commands.add('fireEvent', { prevSubject: true }, (element, event, value)
   element.focus();
   fireEvent[event](element[0], { target: { value } });
 });
-
-// disable screenshots in debug mode. So there is no diffrence to ci.
-if (Cypress.browser.isHeaded) {
-  const noop = () => {};
-  Cypress.Commands.overwrite('matchImageSnapshot', noop);
-}