+++ e2e/tests/settings.js
@@ -7,6 +7,7 @@ export const fixtures = [
   'inline-styles',
   'lists',
   'nested-lists',
+  'aligment_with_punctuations',
   'indent_blocks_lists_with_alignment',
   'quote',
   {
@@ -36,11 +37,8 @@ export const fixtures = [
   'images-sizes',
   'images-original-size',
   {
-    fixture: 'link-preview',
+    fixture: 'link-preview-render',
     config: usePlugins(plugins.all),
-    additionalCommands: cy => {
-      cy.waitForHtmlToLoad();
-    },
   },
 ];