+++ packages/editor/web/src/RichContentEditor/createPlugins.jsx
@@ -2,11 +2,10 @@ import { composeDecorators } from 'draft-js-plugins-editor';
 import createFocusPlugin from 'draft-js-focus-plugin';
 import createResizeDecoration from './Decorators/Resize';
 import createBlockDndPlugin from 'draft-js-drag-n-drop-plugin';
-import { simplePubsub } from 'wix-rich-content-editor-common';
 import createHandleDrop from './handleDrop';
 import createListPlugin from 'draft-js-list-plugin';
 
-const createPlugins = ({ plugins, context }) => {
+const createPlugins = ({ plugins, context, commonPubsub }) => {
   const focusPlugin = createFocusPlugin();
   const resizePlugin = createResizeDecoration({
     horizontal: 'absolute',
@@ -15,7 +14,7 @@ const createPlugins = ({ plugins, context }) => {
     isMobile: context.isMobile,
   });
 
-  const listPlugin = createListPlugin({ olRegex: /1\./, allowNestedLists: false });
+  const listPlugin = createListPlugin({ olRegex: /1\./, allowNestedLists: false, ulChars: [] });
 
   const dndPlugin = createBlockDndPlugin();
   const handleDrop = dndPlugin.handleDrop;
@@ -31,7 +30,7 @@ const createPlugins = ({ plugins, context }) => {
 
   const wixPluginConfig = {
     decorator: wixPluginsDecorators,
-    commonPubsub: simplePubsub(),
+    commonPubsub,
     pluginDefaults,
     ...context,
     ...context.config,
@@ -44,11 +43,15 @@ const createPlugins = ({ plugins, context }) => {
   let pluginTextButtons = [];
   let pluginStyleFns = [];
   wixPlugins.forEach(wixPlugin => {
+    const InsertPluginButtons = wixPlugin.InsertPluginButtons?.map(insertPluginButton => ({
+      ...insertPluginButton,
+      blockType: wixPlugin.blockType,
+    }));
     externalizedButtonProps = [
       ...externalizedButtonProps,
       ...(wixPlugin.externalizedButtonProps || []),
     ];
-    pluginButtons = [...pluginButtons, ...(wixPlugin.InsertPluginButtons || [])];
+    pluginButtons = [...pluginButtons, ...(InsertPluginButtons || [])];
     /* eslint-disable new-cap */
     pluginTextButtons = [
       ...pluginTextButtons,