+++ packages/plugin-image/web/src/toolbar/createToolbar.ts
@@ -1,8 +1,8 @@
 import createInlineButtons from './inline-buttons';
 import createInsertButtons from './insert-buttons';
+import { CreatePluginToolbar } from 'wix-rich-content-common';
 
 const createToolbar: CreatePluginToolbar = ({
-  helpers,
   t,
   anchorTarget,
   relValue,
@@ -19,7 +19,7 @@ const createToolbar: CreatePluginToolbar = ({
       isMobile,
       settings,
     }),
-    InsertButtons: createInsertButtons({ helpers, t, settings }),
+    InsertButtons: createInsertButtons({ t, settings }),
     name: 'image',
   };
 };