+++ packages/ricos-common/web/src/RicosEngine.tsx
@@ -1,15 +1,25 @@
 import React, { Component, Children, FunctionComponent } from 'react';
-import themeStrategy from './themeStrategy/themeStrategy';
+import createThemeStrategy, { ThemeStrategyFunction } from './themeStrategy/themeStrategy';
 import pluginsStrategy from './pluginsStrategy/pluginsStrategy';
 import localeStrategy from './localeStrategy/localeStrategy';
 import { merge } from 'lodash';
 import { isDefined } from 'ts-is-present';
-import './styles.css';
+import previewStrategy from './previewStrategy/previewStrategy';
+import {
+  RicosEditorProps,
+  RicosViewerProps,
+  RichContentChild,
+  RichContentProps,
+  ThemeGeneratorFunction,
+  PreviewSettings,
+} from './types';
 
 interface EngineProps extends RicosEditorProps, RicosViewerProps {
   children: RichContentChild;
   RicosModal: FunctionComponent;
   isViewer: boolean;
+  isPreviewExpanded?: boolean;
+  onPreviewExpand?: PreviewSettings['onPreviewExpand'];
 }
 
 interface EngineState {
@@ -17,9 +27,11 @@ interface EngineState {
 }
 
 export class RicosEngine extends Component<EngineProps, EngineState> {
+  themeStrategy: ThemeStrategyFunction;
   constructor(props: EngineProps) {
     super(props);
     this.state = { localeStrategy: { locale: props.locale } };
+    this.themeStrategy = createThemeStrategy();
   }
 
   static defaultProps = { locale: 'en', isMobile: false };
@@ -42,26 +54,47 @@ export class RicosEngine extends Component<EngineProps, EngineState> {
   }
 
   runStrategies() {
-    const { cssOverride, theme, plugins = [], isViewer = false, content, children } = this.props;
+    const {
+      cssOverride,
+      theme,
+      plugins = [],
+      isViewer = false,
+      content,
+      preview,
+      isPreviewExpanded = false,
+      onPreviewExpand,
+      children,
+    } = this.props;
     const { localeStrategy } = this.state;
 
     const themeGeneratorFunctions: ThemeGeneratorFunction[] = plugins
       .map(plugin => plugin.theme)
       .filter(isDefined);
 
-    const { theme: themeStrategyResult, rawCss } = themeStrategy(
+    const { theme: themeStrategyResult, rawCss } = this.themeStrategy({
       isViewer,
       themeGeneratorFunctions,
-      theme?.palette,
-      cssOverride
+      palette: theme?.palette,
+      cssOverride,
+    });
+
+    const strategyProps = merge(
+      { theme: themeStrategyResult },
+      pluginsStrategy(isViewer, plugins, children.props, themeStrategyResult, content),
+      localeStrategy
+    );
+
+    const { initialState: previewContent, ...previewStrategyResult } = previewStrategy(
+      isViewer,
+      isPreviewExpanded,
+      onPreviewExpand,
+      preview,
+      content
     );
 
     return {
-      strategyProps: merge(
-        { theme: themeStrategyResult },
-        pluginsStrategy(isViewer, plugins, children.props, themeStrategyResult, content),
-        localeStrategy
-      ),
+      strategyProps: merge(strategyProps, previewStrategyResult),
+      previewContent,
       rawCss,
     };
   }
@@ -73,13 +106,14 @@ export class RicosEngine extends Component<EngineProps, EngineState> {
       isMobile,
       toolbarSettings,
       modalSettings = {},
+      isPreviewExpanded,
       placeholder,
       content,
       RicosModal,
       onError,
     } = this.props;
 
-    const { strategyProps, rawCss } = this.runStrategies();
+    const { strategyProps, previewContent, rawCss } = this.runStrategies();
 
     const { useStaticTextToolbar, textToolbarContainer, getToolbarSettings } =
       toolbarSettings || {};
@@ -92,7 +126,7 @@ export class RicosEngine extends Component<EngineProps, EngineState> {
       textToolbarType:
         !isMobile && (textToolbarContainer || useStaticTextToolbar) ? 'static' : 'inline',
       config: { getToolbarSettings },
-      initialState: content,
+      initialState: previewContent || content,
       placeholder,
       onError,
       helpers: {
@@ -106,7 +140,12 @@ export class RicosEngine extends Component<EngineProps, EngineState> {
       <style type="text/css" key={'styleElement'}>
         {rawCss}
       </style>,
-      <RicosModal ariaHiddenId={ariaHiddenId} {...mergedRCProps} key={'ricosElement'}>
+      <RicosModal
+        ariaHiddenId={ariaHiddenId}
+        isModalSuspended={previewContent && !isPreviewExpanded}
+        {...mergedRCProps}
+        key={'ricosElement'}
+      >
         {Children.only(React.cloneElement(children, { ...mergedRCProps }))}
       </RicosModal>,
     ];