+++ packages/ricos-common/web/src/RicosTypes.ts
@@ -1,4 +1,19 @@
-interface RichContentProps {
+import { Decorator, Helpers, GetToolbarSettings } from 'wix-rich-content-common';
+import { EditorState } from 'draft-js';
+import { ReactElement } from 'react';
+import {
+  RicosContent,
+  RicosCssOverride,
+  Palette,
+  PalettePreset,
+  InlineStyleMapper,
+  ModalsMap,
+  PluginConfig,
+  TypeMapper,
+  PreviewSettings,
+} from './types';
+
+export interface RichContentProps {
   config?: Record<string, unknown>;
   decorators?: Decorator[];
   editorKey?: string;
@@ -9,21 +24,23 @@ interface RichContentProps {
   locale?: string;
   localeResource?: Record<string, unknown>;
   ModalsMap?: ModalsMap;
-  onChange?(editorState: DraftEditorState): void;
+  onChange?(editorState: EditorState): void;
   onError?: OnErrorFunction;
   placeholder?: string;
   plugins?: PluginConfig[];
   textToolbarType?: TextToolbarType;
   theme?: RicosCssOverride;
   typeMappers?: TypeMapper[];
+  transformation?: Record<string, unknown>;
+  seoMode?: boolean | Record<string, unknown>;
 }
 
-interface ExportedRichContentProps extends RichContentProps {
+export interface ExportedRichContentProps extends RichContentProps {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   [propName: string]: any;
 }
 
-interface RicosProps {
+export interface RicosProps {
   _rcProps?: RichContentProps; // For internal use by WixRicos only
   children?: RichContentChild;
   content?: RicosContent;
@@ -35,40 +52,42 @@ interface RicosProps {
   theme?: RicosTheme;
 }
 
-interface RicosEditorProps extends RicosProps {
+export interface RicosEditorProps extends RicosProps {
   modalSettings?: ModalSettings;
   onChange?: OnContentChangeFunction;
   placeholder?: string;
   toolbarSettings?: ToolbarSettings;
 }
 
-type RicosViewerProps = RicosProps;
+export interface RicosViewerProps extends RicosProps {
+  preview?: PreviewSettings;
+}
 
-interface RicosTheme {
+export interface RicosTheme {
   palette?: Palette | PalettePreset;
 }
 
-type RichContentChild = import('react').ReactElement<ExportedRichContentProps>;
+export type RichContentChild = ReactElement<ExportedRichContentProps>;
 
-interface ModalSettings {
+export interface ModalSettings {
   openModal?: (data: Record<string, unknown>) => void;
   closeModal?: () => void;
   ariaHiddenId?: string;
 }
 
-type TextToolbarType = 'inline' | 'static';
+export type TextToolbarType = 'inline' | 'static';
 
-interface ToolbarSettings {
+export interface ToolbarSettings {
   getToolbarSettings?: GetToolbarSettings;
   textToolbarContainer?: HTMLElement;
   useStaticTextToolbar?: boolean;
 }
 
-interface EditorDataInstance {
+export interface EditorDataInstance {
   getContentState: () => RicosContent;
-  refresh: (editorState: DraftEditorState) => void;
+  refresh: (editorState: EditorState) => void;
 }
 
-type OnContentChangeFunction = (content: RicosContent) => void;
+export type OnContentChangeFunction = (content: RicosContent) => void;
 
-type OnErrorFunction = (error: string) => void;
+export type OnErrorFunction = (error: string) => void;